Add array data sources.
Array data sources are useful for data where the whole array might change at once but individual elements never change. Using diffable data sources is difficult in this case because the element identifier is its index, and in that case, changing the whole array does not trigger an update for elements where the indices stay the same. For example, the first element of the previous array and the first element of the new value array will appear to be the same element to a diffable data source if its id is its index.

It's possible to model this with a diffable data source if each array element has a UUID as its identifier. But then we have to store those UUIDs with the elements. It's a more complicated solution.